Remove function to replace infinities from the algorithm. Now it should be called by user. Also added one more useful util function which validate input data.
