+++ tests/munkres_test.cpp
@@ -95,136 +95,6 @@ TYPED_TEST_CASE (MunkresTestFloatingLong, TypesFloatingLong);
 
 
 
-TYPED_TEST (MunkresTestFloating, replace_infinites_4x4Case001_Success)
-{
-    // Arrange.
-    const auto infinity = std::numeric_limits<TypeParam>::infinity ();
-    Matrix<TypeParam> etalon_matrix {
-        { 1.0,      0.0,      3.0,      2.0},
-        { 3.0,     -2.0,     -1.0,      0.0},
-        {-1.0,      3.0,      2.0,      0.0},
-        {-1.0,      0.0,      2.0,      3.0}
-    };
-    Matrix<TypeParam> test_matrix {
-        { 1.0,      0.0,     infinity,  2.0l},
-        {infinity, -2.0,     -1.0l,     0.0l},
-        {-1.0,     infinity,  2.0l,     0.0l},
-        {-1.0,      0.0,      2.0l, infinity}
-    };
-
-    // Act.
-    munkres_cpp::replace_infinites (test_matrix);
-
-    // Assert.
-    EXPECT_EQ (etalon_matrix, test_matrix);
-}
-
-
-
-TYPED_TEST (MunkresTestFloating, replace_infinites_4x4Case002_Success)
-{
-    // Arrange.
-    const auto infinity = std::numeric_limits<TypeParam>::infinity ();
-    Matrix<TypeParam> etalon_matrix {
-        { 3.0,      0.0,      3.0,      2.0},
-        { 3.0,     -2.0,     -1.0,      3.0},
-        {-1.0,      3.0,      2.0,      3.0},
-        {-1.0,      3.0,      2.0,      3.0}
-    };
-    Matrix<TypeParam> test_matrix {
-        {infinity,  0.0,     infinity,  2.0},
-        {infinity, -2.0,     -1.0, infinity},
-        {-1.0,     infinity,  2.0, infinity},
-        {-1.0,     infinity,  2.0, infinity}
-    };
-
-    // Act.
-    munkres_cpp::replace_infinites (test_matrix);
-
-    // Assert.
-    EXPECT_EQ (etalon_matrix, test_matrix);
-}
-
-
-
-TYPED_TEST (MunkresTestFloating, replace_infinites_4x4Case003_Success)
-{
-    // Arrange.
-    const auto infinity = std::numeric_limits<TypeParam>::infinity ();
-    Matrix<TypeParam> etalon_matrix {
-        {-5.0,     -4.0,     -1.0,     -2.0},
-        {-1.0,     -2.0,     -5.0,     -4.0},
-        {-5.0,     -1.0,     -2.0,     -4.0},
-        {-5.0,     -4.0,     -2.0,     -1.0}
-    };
-    Matrix<TypeParam> test_matrix {
-        {-5.0,     -4.0,     infinity, -2.0},
-        {infinity, -2.0,     -5.0,     -4.0},
-        {-5.0,     infinity, -2.0,     -4.0},
-        {-5.0,     -4.0,     -2.0, infinity}
-    };
-
-    // Act.
-    munkres_cpp::replace_infinites (test_matrix);
-
-    // Assert.
-    EXPECT_EQ (etalon_matrix, test_matrix);
-}
-
-
-
-TYPED_TEST (MunkresTestFloating, replace_infinites_4x4Case004_Success)
-{
-    // Arrange.
-    const auto infinity = std::numeric_limits<TypeParam>::infinity ();
-    Matrix<TypeParam> etalon_matrix {
-        { 1.0,      0.0,      3.0,      2.0},
-        { 3.0,     -2.0,     -1.0,      0.0},
-        {-1.0,      3.0,      0.0,      0.0},
-        {-1.0,      0.0,      0.0,      3.0}
-    };
-    Matrix<TypeParam> test_matrix {
-        { 1.0,      0.0,     infinity,  2.0},
-        {infinity, -2.0,     -1.0,      0.0},
-        {-1.0,     infinity,  0.0,      0.0},
-        {-1.0,      0.0,      0.0, infinity}
-    };
-
-    // Act.
-    munkres_cpp::replace_infinites (test_matrix);
-
-    // Assert.
-    EXPECT_EQ (etalon_matrix, test_matrix);
-}
-
-
-
-TYPED_TEST (MunkresTestFloating, replace_infinites_4x4Case005_Success)
-{
-    // Arrange.
-    const auto infinity = std::numeric_limits<TypeParam>::infinity ();
-    Matrix<TypeParam> etalon_matrix {
-        {     0.0,      0.0,      0.0,      0.0},
-        {     0.0,      0.0,      0.0,      0.0},
-        {     0.0,      0.0,      0.0,      0.0},
-        {     0.0,      0.0,      0.0,      0.0}
-    };
-    Matrix<TypeParam> test_matrix {
-        {infinity, infinity, infinity, infinity},
-        {infinity, infinity, infinity, infinity},
-        {infinity, infinity, infinity, infinity},
-        {infinity, infinity, infinity, infinity}
-    };
-
-    // Act.
-    munkres_cpp::replace_infinites (test_matrix);
-
-    // Assert.
-    EXPECT_EQ (etalon_matrix, test_matrix);
-}
-
-
-
 TYPED_TEST (MunkresTestAll, minimize_along_direction_5x5_OverRowsOnly_Success)
 {
     // Arrange.
